ASKAPSDP-1659 - Refactoring to allow catalogue type.
CASDA have a new specification that the type of catalogue needs to be 
given. The code in CasdaUploadApp was not quite flexible enough to 
allow that to be easily dropped in, so I've refactored the "Element" 
classes to have several base classes, depending on whether or not they 
need a project and type entry. This simplifies the interface that the  
buildArtifactElements function has to deal with.