ASKAPSDP-1668 - Initial fix for the width & precision in the CASDA catalogues.
This commit changes the way the catalogues are processed, so that only 
the widths of the string columns are changed prior to writing to the 
XML/VOtable file. All other columns have their widths & precisions 
left as the initial values. Before writing the to ASCII files, the 
full checking is done, potentially altering the precision & widths of 
all columns.
This keeps the VOTables for the Component & Island catalogues matching 
the spec.