ASKAPSDP-1659 - Correcting the MS filename in the XML file, and adding robustness to provided parameters.
The earlier changes to the writing of the XML entries removed the .tar 
suffix from the filename entry for MSs. This change replaces it (since 
the XML file needs to match the files in that directory). 
The change also checks for the use of artefactlist instead of 
artifactlist and gives a warning since both are actually valid ways to 
spell artifact!