ASKAPSDP-1675 - Improvements to Absorption objects & catalogues
The date and frequency values for an absorption object are calculated 
better (the latter still not correctly, but at least you get a fixed 
value now). The interface to the catalogue has changed as well, so 
that a vector of component/object pairs is provided, rather than a 
multimap (components can't be ordered as yet).