ASKAPSDP-1477 - Adding restoring beam handling to simager
This change allows the restoring beams of each worker to be passed 
back to the master, who records them in a map (tagged to the global 
channel number). The restoring beam is also written to the restored 
image, using the new beamReference input parameter to determine which 
channel is used (this has the same interface as the old makecube).