ASKAPSD-1874 - Fixing name-parsing bug.
This change fixes a bug in processing the image name. Previously, if 
the image name did not have a '.' in it, the processing would crash. 
It now searches for one first.
Also improves the use of the sbid, whereby this is put in the ID 
string only if it is not 'null' (ie. the user has provided it in the 
parset).