ASKAPSDP-2012 - Archiving pipeline improvements to better handle thumbnails.
This change alters the scripts to take advantage of the new ability of 
casdaupload to specify thumbnails at the image element level in the 
observation.xml file. The way the getArtifactsLists script works is 
simplified a bit, keeping the 2D and non-2D image lists separate. 
The way the thumbnail sizes are specified has changed slightly 
(suffixes are now '_large' & '_small').