ASKAPSDP-2018 - Using FITS files in source-finding pipeline scripts
This change makes the source-finding script use FITS versions of the 
images. It incorporates the convertToFITS functionality at the start 
of the selavy job, by extracting that code out and putting it in a 
function defined in utils.h. This function provides the text that 
allows it to be put into a slurm file. 
The convertToFITS script now makes use of this function (so we only 
have the code in one place).
It also now just uses the pyfits library, rather than getting it out 
of astropy.io.fits (since pyfits is part of the askapsoft module now).