ASKAPSDP-1593 - Overhauling the Selavy extraction code
This change improves the extraction code in Selavy, allowing the 
extraction of spectra for fitted components from cubes that may have 
a different pixel grid. The pixel localisation is based on the WCS 
position. 
This change also cleans up the code a little, fixing a few bugs that 
prevented things working well particularly with FITS file input.