ASKAPSDP-1593 - Further extraction fixes
This change updates the extraction code to better handle the 
noise-spectrum extraction. This now has the same behaviour of being 
able to extract noise spectra for individual fitted components. 
The unit tests have also been updated (as they were failing after the 
previous commit), with a couple of necessary changes to the main code.