ASKAPSDP-2061 - Fixing import of robust_contsub
This change ensures that we find the robust_contsub module, and import 
it correctly. Since it is located in $ACES/tools, we can't assume it 
is in the user's python path, so we find the directory and explicitly 
load it. If we don't find it, we don't run the task.