ASKAPSDP-2125, ASKAPSDP-2176 - improved flagging dependencies
This change adds a separate job to do the flagging on the averaged 
dataset, as the dependencies were all wrong when I tried to use the 
same script twice. This now introduces a new set of flagging 
parameters for the averaged-data job.
Have also tidied up the dependencies for subsequent jobs.