ASKAPSDP-2205 - improving bandpass calibration in the pipeline.
This change forces the processing of the bandpass calibrator to handle 
all beams up to the requested maximum, even if BEAM_MIN>0. This way 
they will all be passed to cbpcalibrator, and it will work OK. 
Have also improved the handling of the bandpass table name - if a 
leading path is given, it is left in that location, else it is changed 
to lie in the BPCAL directory. Also improving the run-time output and 
the documentation.