ASKAPSDP-1991 - SB transition upon CASDA ingest
This change adds a slurm job to the end of the archiving scripts, that 
polls the CASDA upload directory for the DONE file written by CASDA 
once ingest is complete. 
Additional user parameters added to define the polling cycle, and the 
time out period (where we give up looking for it and raise an error).
Documentation updated.