ASKAPSDP-2174 - More flexibility in self-calibration parameters
This change allows the user to vary the major cycle threshold and the 
number of major cycles with each self-calibration loop. This entailed 
a new function in utils that is called from within the slurm job. We 
set the arrays in the same way as for the SELFCAL_INTERVAL etc 
parameters.
Also introducing the ability to give the minUV distance for imaging 
and calibration - this too can vary with self-cal loop number.