ASKAPSDP-2338 - Fixing parsing of cube names with %p
This change fixes issues in parsing the names of cubes that have the 
%p wildcard in them. This applies to both the general extraction case 
and the specific implementation in the polarisation data.
The beam scale factors had to be turned into maps (with the Stokes 
parameter as key), and we now store the input image units as part of 
the SourceDataExtractor class (removing the need to open the input 
image when writing).