ASKAPSDP-2339 - Additional parameterisation of RM synthesis results.
This change adds the last remaining parameters for the polarisation 
catalogue - the edge flag and the two complexity parameters. It also 
slightly improves the debiasing, only doing it when the SNR is above 
the debiasing threshold. 
The RMSynthesis interface now allows access to the input fractional 
polarisation spectrum and the lambda-squared array (to make the 
parameterisation simpler).