ASKAPSDP-2395 - Enabling export of the FDF
This change adds a new class, FDFwriter, that allows straightforward 
writing of the FDF and RMSF arrays to 1D complex CASA images. This is 
implemented from within the definition of the CasdaPolarisationEntry, 
just after the Stokes spectra are written.
This change also sees a few minor changes to class interfaces, and a 
bit of tidying up and documenting of some classes.