ASKAPSDP-2366 - better management of selfcal/imaging loop parameters
This change improves the way processDefaults parses and checks the 
loop-enabled parameters for imaging & self-cal. It will only check 
them if the DO_CONT_IMAGING flag is true, and only check the selfcal 
parameters if DO_SELFCAL is true. 
Have also removed the case where we reset the SELFCAL_NUM_LOOPS to 
zero for non-selfcal imaging - this was unncessary, and was resulting 
in confusing error messages.