ASKAPSDP-2431 - Addition of RM Synthesis to continuum source-finding
This change adds the RM Synthesis parameters to the continuum 
source-finding job. This will only be done for the regular beam & 
mosaic images (not the loop mosaics), and only if the DO_RM_SYNTHESIS 
flag is true. 
This has entailed renaming sourcefinding.sh to sourcefindingCont.sh 
for clarity, and expanded its code somewhat to also account for the 
continuum cubes. 
Have also moved the determination of the imageName etc for the 
source-finding jobs (spectral-line included) into the individual 
scripts and away from the scripts that are calling them.