ASKAPSDP-2433 - Different sourcefinding switches.
This introduces separate switches for the continuum & spectral-line 
sourcefinding, so that we can have one without the other. They default 
to "", and if not specified take on the value of the relevant imaging 
switch (so if you are doing continuum imaging, you don't need to worry 
about the spectral-line sourcefinding at all).