ASKAPSDP-2414 - Better tests at start of linmos scripts.
This change adds better tests for the existence of mosaicked images 
that are run at the start of the linmos jobs. The tests look for each 
output file created by the linmos task, and if any of them exist (when 
CLOBBER=false) then we don't run the script. The tests match the loops 
that occur within the slurm job.