ASKAPSDP-2433 - Improving source-finding job dependencies
This change address the case where there is only a single field, so we 
don't make the multi-field mosaic. In this case, the source-finding 
jobs for this case would still run, but without dependencies, so you'd 
get an orphan job running at the start not doing anything.
Now, if it is the top field (FIELD=".") and there are no dependencies, 
we don't run.