Fixes for CASDA upload for ASKAPSDP-2472 and ASKAPSDP-2480
This ensures that the wildcards presented to casdaupload are suitable 
to uniquely identify each spectrum in the uplaod directory.

We also do not upload the spectral weights image - there is no 
corresponding image type, so trying to upload it will result in an 
error.

The image types for the other weights images have been changed to 
cont_sensitivity_2d/3d, as there is no explicit _weights_ type.