ASKAPSDP-2520 - Improving use of the new imager in the pipelines.
This changes a few parameter names, and improves the way we interact 
with the subcubes produced by the new imager.
New parameters are:
* NUM_SPECTRAL_CUBES for NSUB_CUBES
* DO_BARY now defaults to true
* wrList is now SUBBAND_WRITER_LIST, which is calculated in 
processDefaults in the proper way

The spectral mosaicking scripts linmosSpectral & linmosFieldsSpectral 
now incorporate a loop over the subbands, removing the need for the 
additional altLinmos (it has been removed from scienceCalIm).

The setImageProperties function now knows about the subband and 
NUM_SPECTRAL_CUBES, and can form the image names accordingly.