ASKAPSDP-2520 - Improvements to imcontsub script
This adds a loop over the subbands for the image-based 
continuum-subtraction, so that it is able to handle multiple sub cubes 
produced by the new imager.
Makes use of the new capabilities of the setImageProperties function.
Also tidied up the python script and the bash code at the same time.