ASKAPSDP-2527 - Stopping the pipeline re-running the split job
If we have previously run the pipeline and purged the full-resolution 
MS, then on subsequent runs we don't want to re-do the splitting to 
create another full-resolution dataset (as this then won't be purged 
and we have no need of it). This change turns off splitting if the 
PURGE flag is on and the averaged dataset exists.