ASKAPSDP-2670 - Better FITS-handling for pipeline.
This change allows the setImageProperties function to add the .fits 
extension if the relevant IMAGETYPE is set to "fits". 
It undoes several of the recent changes, as they will no longer be 
required.
The defineArtifactsScript is updated to handle this addition, when 
ADD_FITS_SUFFIX is set to true, the .fits extension is only added when 
the IMAGETYPE is not "fits" (since it will already be there). 
This does not apply to extracted spectra etc from Selavy - these are 
not currently written to FITS by Selavy.
The convertToFITS script also has a check to see if a fits extension 
is present, and only does the conversion if it is not.