ASKAPSDP-2473 - Allowing distributed parameterisation of sources
This is a merge from features/selavy-dev.
It allows distributed parameterisation of:
* RM synthesis on components
* HI emission-line objects
* Absorption-line objects (not fully implemented as a mode of selavy 
yet though).
This gives noticeable speed-ups in the processing by parallelising the 
parameterisation & extraction associated with individual sources.
Also includes a few fixes to issues with HI parameters in particular, 
especially the determination of local noise for distributed 
RadioSource objects.