ASKAPSDP-2721 - Fixing a bug introduced earlier with the alpha/beta columns in the VOTables.
They were getting very large precisions, but this was due to the 
secondary components not being initialised properly.
Have also fixed an issue with the width of columns in the VOTable, 
where we were widening to take into account the ASCII header & units - 
we don't need to do that with VOTables.