ASKAPSDP-2762 - Fixing class structure for FitCatalogue, as well as handling of the fitType parameter.
This change ensures we have correct constructors for the different 
uses of ComponentCatalogue, allowing the specification of a fit type 
and also allowing it to default to the preferred CASDA type. This 
allows different FitCatalogues to be created for different types.

It also fixes a bug where the incorrect number of fits was being used 
when creating the individual components - this is now taken from the 
correct fit type.