ASKAPSDP-2846 - Adding SNR flux limit for components in self-cal
This adds the ability in the self-calibration to reject components 
that lie below some signal-to-noise level. The average image noise is 
found from the selavy catalogue, and converted to a flux level via the 
input SNR limit parameter. This is then given to cmodel to impose the 
flux limit.
Defaults and documentation updated.