Merging from ASKAPSDP-2591--selavy-fitting feature branch, containing fixes for ASKAPSDP-2591, ASKAPSDP-2971 and ASKAPSDP-2983.
Selavy now records the best fit, even if it is formally bad according 
to chi-squared. It can reduce the number of Gaussians being fit if the 
fit fails to converge. And the curvature-map calculations will now 
take into account the weight images in the same was as the rest of 
Selavy.