ASKAPSDP-3127 - Improvements to Selavy's HI catalogue.
This gives valid default/initial values for all parameters in the 
Selavy HI catalogue, and fixes a bug in determining the fitted 
major/minor axes, where the mask was not being taken into account. 
This was resulting in nan values being returned. We now construct a 
MaskedArray from which to find the maximum (peak) flux.