ASKAPSDP-3157 - Further fix to get the paths for MS artifacts correct.
I had changed one instance of the code, but not the bit that went in 
the XML file. This is now done in the same way as all other elements.

Also a few changes to the main program - adding a try/catch for any 
errors (so they get reported in a standard way), and reporting the 
package version at the start of the log. Other minor edits to log 
messages as well.