ASKAPSDP-3178 - Making the Fitter::acceptable() only operate when explicitly requested.
This introduces a fitting parameter, Fitter.applyAcceptanceCriteria, 
that turns on the usual use of Fitter::acceptable() - it now returns 
the value of the passChisq() test if that parameter is set to false 
(its default). The other method, Fitter::acceptableExceptChisq() will 
return true always unless this parameter is set to true.