ASKAPSDP-2124 - Initial version of merging & splitting to handle frequency-split raw data
This adds options in the preparation of the data for both the bandpass 
calibration and the science observations, so that we use msmerge, or 
mssplit+msmerge, to produce the input MS used by the other pipeline 
tasks. These methods are implemented when there are more MSs than 
beams in the observation.
The calculation of channel numbers are not handled properly yet.