ASKAPSDP-3393 - Adding better checks for negative flux components
This changes the approach for dealing with fitted components that have 
negative peak fluxes. Previously these were set to zero flux, but now 
we reject the fit (unless the negativeFluxPossible flag is set).

The handling of this flag is also improved - it was never properly initialised.

The value of the negative-flux test is reported in the same way as the 
other tests - it forms one of the basic tests (that is always done).