ASKAPSDP-3289 - Manual merge of current state of the features/ASKAPSDP-3289 branch to trunk.
This will enable the split-by-time mode in the trunk version of the pipeline scripts.
This had to be done by manually copying the files that had changed in the pipelines bin directory and in the documentation directory. 

The reintegrate merge failed with an error like this:
/group/askap/whi550/askapsdp-clean-trunk> svn merge --dry-run ^/features/ASKAPSDP-3289
svn: E195016: Reintegrate can only be used if revisions 10328 through 10658 were previously merged from https://svn.atnf.csiro.au/askapsdp/trunk to the reintegrate source, but this is not the case:
  features/ASKAPSDP-3289/Code/Components/Synthesis/synthesis/current/distributedimager
    Missing ranges: /trunk/Code/Components/Synthesis/synthesis/current/distributedimager:10367
  features/ASKAPSDP-3289/Code/Components/Synthesis/synthesis/current/functests/continuum
    Missing ranges: /trunk/Code/Components/Synthesis/synthesis/current/functests/continuum:10367
  features/ASKAPSDP-3289/Code/Components/Synthesis/synthesis/current/functests/spectralline
    Missing ranges: /trunk/Code/Components/Synthesis/synthesis/current/functests/spectralline:10367
  features/ASKAPSDP-3289/Code/Components/Synthesis/synthesis/current/messages
    Missing ranges: /trunk/Code/Components/Synthesis/synthesis/current/messages:10367