ASKAPSDP-3560 - Unsetting LOOP after the selfcal in the new mode.
An effect of moving the LOOP loop to the pipeline script (instead of 
the slurm script) was that $LOOP was defined after the selfcal script 
had run. This mucked up the beam-wise source-finding that was done - 
the image being searched for when 2 loops were being done was a LOOP3 
image!
Have unset LOOP and so that shouldn't happen.