ASKAPSDP-3680 - Implementing removebeam option in pipeline mosaicking.
The continuum mosaicking now has the removebeam option set, and takes 
the number of taylor terms in a consistent way.

Because the taylor.1 will be mosaicked at the same time as the 
taylor.0, we remove the loop over taylor terms and just set TTERM=0. 
This will reduce the number of linmos jobs!