ASKAPSDP-3754 - Refactoring, and fixing the width of columns in VOTable output.
The main effect of this change is to enforce the CASDA limits to the column widths in the VOTable output.

To support this and reduce code duplication, I've introduced base 
classes in catalogues/ - we now have CasdaObject and CasdaCatalogue to 
form the base classes for the object classes and the catalogue classes 
respectively. This has removed a lot of common code, particularly from 
the catalogue classes.