Simplifying the slurm dependencies used in the imaging within the pipeline.
This defines the set of pre-imaging job IDs to be used as dependencies for the imaging jobs and those following.
This sets it once in a central place, and allows adaptation with further changes to come (adapting the flag summary scripts from ASKAPSDP-3766 to be used with the rapid survey scripts for ASKAPSDP-3495).