Rapid survey processing option for pipeline. This addresses the following tickets:
ASKAPSDP-3492 - there is a new script created that runs the pre-imaging tasks for the continuum processing: splitting, bandpass application, flagging, averaging, flagging.
ASKAPSDP-3494 - this script is called at the start of each of the continuum imaging jobs, so that we only get a single slurm job to do everything up to the continuum image.
ASKAPSDP-3495 - implementation of this in a "rapid" version of scienceCalIm. This is triggered by the new DO_RAPID_SURVEY config parameter. No spectral processing is done in this mode.