ASKAPSDP-3755 - Further fix to modules for getMatchingMS.py
This ensures we use the correct python when running getMatchingMS.py in the pipeline scripts.

We need to unload the askapsoft module, load numpy, and then load 
askapsoft again. If the user has not loaded python, then loading numpy 
will load the python module as well, meaning it will provide the 
python executable and it won't know about python-casacore.

This change ensures the python will be taken from the askapsoft module.